:recycle: Format: 알림 및 하우스메이트 모달 코드 리팩토링
