+++ src/components/common/Modal/hooks/useClickOutside.ts
@@ -1,0 +1,33 @@
+import { useEffect } from 'react';
+
+interface UseClickOutsideProps {
+  modalRef: React.RefObject<HTMLDivElement>;
+  buttonRef: React.RefObject<HTMLButtonElement>;
+  isOpen: boolean;
+  onClose: () => void;
+}
+
+export const useClickOutside = ({
+  modalRef,
+  buttonRef,
+  isOpen,
+  onClose,
+}: UseClickOutsideProps) => {
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        isOpen &&
+        modalRef.current &&
+        !modalRef.current.contains(event.target as Node) &&
+        !buttonRef.current?.contains(event.target as Node)
+      ) {
+        onClose();
+      }
+    };
+
+    document.addEventListener('mousedown', handleClickOutside);
+    return () => {
+      document.removeEventListener('mousedown', handleClickOutside);
+    };
+  }, [isOpen, onClose, modalRef, buttonRef]);
+};
