+++ src/components/header/menus/modal/LoadingState.tsx
@@ -1,0 +1,3 @@
+export const LoadingState = () => {
+  return <div className='text-center text-[#3E507D]/70'>로딩 중...</div>;
+};
